Replace index.native with Platform.select
