+++ lib/storage/NativeStorage.js
