+++ lib/storage/__mocks__/index.js
@@ -1,0 +1,3 @@
+import WebStorage from '../WebStorage';
+
+export default WebStorage;
